+++ tests/testthat/test-merge-estimate.R
@@ -1,3 +1,5 @@
+context("merge.estimate")
+
 library(testthat)
 library(lava)